Solution to Manage Book Library

+++ Manage Book Library@@67877d61dc107486ad5c274c/code.c
@@ -1,18 +1,20 @@
 #include <stdio.h>
 
-struct Book {
-    int price;
+struct Book 
+{
+    float price;
     char title[100], author[100];
 };
 
-int main() {
+int main() 
+{
     int n;
     scanf("%d", &n);
 
     struct Book books[n];
     for (int i = 0; i < n; i++) 
     {
-        scanf("%s %s %d", books[i].title, books[i].author, &books[i].price);
+        scanf("%s %s %f", books[i].title, books[i].author, &books[i].price);
     }
 
     float p;
